+++ tsup.config.ts
@@ -5,7 +5,7 @@ export default defineConfig({
         index: "src/index.ts",
         react: "src/plugins/kernel/react/index.tsx",
         testing: "src/testing.ts",
-        server: "src/plugins/seo/react/server.tsx",
+        server: "src/plugins/server/react/server.tsx",
     },
     format: ["esm"],
     dts: true,
@@ -14,7 +14,7 @@ export default defineConfig({
     treeshake: true,
     target: "es2022",
 
-    external: ["react", "react/jsx-runtime", "react-dom", "react-dom/server", "zod", "node:fs", "node:fs/promises", "node:path"],
+    external: ["react", "react/jsx-runtime", "react-dom", "react-dom/server", "zod", "node:fs", "node:fs/promises", "node:path", "node:url", "vite"],
 
     platform: "node",
 });